+++ .next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"Ge4cX49wnJavJV7A9f3NGmPktZ58hmpB7ZRI6JsYJEk=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"wKfZWO42kgV1e1hG/26A1jfhGIG/m3JdbXULVaGFiww=\"\n}"