+++ src/app/dm3assets/page.tsx
@@ -1,6 +1,6 @@
 "use client";
-import DM6AssetsHeroLayout from '@/components/DM6AssetsHeroLayout';
+import DM3AssetsHeroLayout from '@/components/DM3AssetsHeroLayout';
 
 export default function DM3Assets() {
-  return <DM6AssetsHeroLayout />;
+  return <DM3AssetsHeroLayout />;
 }